+++ Source/EventFlow/Exceptions/DuplicateOperationException.cs
@@ -20,36 +20,22 @@
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-using System.Collections.Generic;
-using System.Threading;
-using System.Threading.Tasks;
+using System;
 using EventFlow.Core;
-using EventFlow.EventStores;
 
-namespace EventFlow.Aggregates
+namespace EventFlow.Exceptions
 {
-    public interface IAggregateRoot
+    public class DuplicateOperationException : Exception
     {
-        IAggregateName Name { get; }
-        int Version { get; }
-        IEnumerable<IAggregateEvent> UncommittedEvents { get; }
-        bool IsNew { get; }
+        public ISourceId SourceId { get; }
+        public IIdentity AggregateId { get; }
 
-        Task<IReadOnlyCollection<IDomainEvent>> CommitAsync(
-            IEventStore eventStore,
-            ISourceId sourceId,
-            CancellationToken cancellationToken);
-
-        bool HasSourceId(ISourceId sourceId);
-
-        void ApplyEvents(IEnumerable<IAggregateEvent> aggregateEvents);
-
-        void ApplyEvents(IReadOnlyCollection<IDomainEvent> domainEvents);
-    }
-
-    public interface IAggregateRoot<out TIdentity> : IAggregateRoot
-        where TIdentity : IIdentity
-    {
-        TIdentity Id { get; }
+        public DuplicateOperationException(
+            ISourceId sourceId, IIdentity aggregateId, string message)
+            : base(message)
+        {
+            SourceId = sourceId;
+            AggregateId = aggregateId;
+        }
     }
 }