+++ RELEASE_NOTES.md
@@ -1,4 +1,8 @@
-### New in 0.12 (not released yet)
+### New in 0.13 (not released yet)
+
+ * _Nothing yet_
+
+### New in 0.12.891 (released 2015-09-04)
 
  * Breaking: Aggregate root no longer have `Aggregate` removed from their
    when name, i.e., the metadata property with key `aggregate_name` (or
@@ -22,8 +26,10 @@
    `CommandId` each time the a `Command<,>` instance is created. You can pass
    specific value, merely use the newly added constructor taking the ID.
    Alternatively you commands could inherit from the new
-   `DeterministicIdCommand`, enabling commands with the same state to have the
+   `DistinctCommand`, enabling commands with the same state to have the
    same `SourceId`
+ * New: Duplicate commands can be detected using the new `ISourceId`. Read the
+   EventFlow article regarding commands for more details
  * New: Aggregate names can now be configured using the attribute
    `AggregateName`. The name can be accessed using the new `IAggregateRoot.Name`
    property